Fixes bug where AGF2 two-body energy did not pass chempot to gf

+++ auxgf/hf/hf.py
@@ -154,7 +154,7 @@ class HF:
 
     @property
     def occ(self):
-        return np.stack(self._pyscf.mo_occ)
+        return np.stack(self._pyscf.mo_occ).astype(types.int32)
 
     @property
     def c(self):
